fix dtype of cumsum

+++ src/transformers/models/informer/modeling_informer.py
@@ -647,7 +647,7 @@ class InformerProbSparseAttention(nn.Module):
         # calculate context for updating the attn_output, based on:
         # https://github.com/zhouhaoyi/Informer2020/blob/ac59c7447135473fb2aafeafe94395f884d5c7a5/models/attn.py#L74
         if self.is_decoder:
-            context = value_states.cumsum(dim=-2)
+            context = value_states.cumsum(dim=-2, dtype=torch.float32).to(value_states.dtype)
         else:
             v_mean_dim_time = value_states.mean(dim=-2)
             context = (
